Больше подробностей (часть 2) fix
