+++ src/components/Tab.js
@@ -1,29 +1,11 @@
 import React, {useState} from 'react';
-import PropTypes from "prop-types";
+import PropTypes from 'prop-types';
+import TabOverview from './TabOverview';
+import TabDetails from "./TabDetails";
+import TabReviews from "./TabReviews";
 
 
 const tabs = [`Overview`, `Details`, `Reviews`];
-const rating = {
-  'Bad': [0, 3],
-  'Normal': [3, 6],
-  'Good': [6, 8],
-  'Very good': [8, 10]
-};
-
-const getRatingString = (ratingFilm) => {
-  for (let prop in rating) {
-    if ((ratingFilm >= rating[prop][0]) && (ratingFilm < rating[prop][1])) {
-      return prop;
-    }
-  }
-  return `Awesome`;
-};
-
-const getTimeFromMins = (mins) => {
-  let hours = Math.trunc(mins / 60);
-  let minutes = mins % 60;
-  return hours + `h ` + minutes + `m`;
-};
 
 const Tab = ({film = {}}) => {
 
@@ -32,167 +14,13 @@ const Tab = ({film = {}}) => {
   const getComponentByType = (tabFilm) => {
     switch (tabFilm) {
       case `Details`:
-        return <>
-          <div className="movie-card__text movie-card__row">
-            <div className="movie-card__text-col">
-              <p className="movie-card__details-item">
-                <strong className="movie-card__details-name">Director</strong>
-                <span className="movie-card__details-value">{film.director}</span>
-              </p>
-              <p className="movie-card__details-item">
-                <strong className="movie-card__details-name">Starring</strong>
-                <span className="movie-card__details-value">
-                    Bill Murray, <br/>
-                    Edward Norton, <br/>
-                    Jude Law, <br/>
-                    Willem Dafoe, <br/>
-                    Saoirse Ronan, <br/>
-                    Tony Revoloru, <br/>
-                    Tilda Swinton, <br/>
-                    Tom Wilkinson, <br/>
-                    Owen Wilkinson, <br/>
-                    Adrien Brody, <br/>
-                    Ralph Fiennes, <br/>
-                    Jeff Goldblum
-                </span>
-              </p>
-            </div>
-
-            <div className="movie-card__text-col">
-              <p className="movie-card__details-item">
-                <strong className="movie-card__details-name">Run Time</strong>
-                <span className="movie-card__details-value">{getTimeFromMins(film.run_time)}</span>
-              </p>
-              <p className="movie-card__details-item">
-                <strong className="movie-card__details-name">Genre</strong>
-                <span className="movie-card__details-value">{film.genre}</span>
-              </p>
-              <p className="movie-card__details-item">
-                <strong className="movie-card__details-name">Released</strong>
-                <span className="movie-card__details-value">{film.released}</span>
-              </p>
-            </div>
-          </div>
-        </>;
+        return <TabDetails film={film} />;
 
       case `Reviews`:
-        return <>
-          <div className="movie-card__reviews movie-card__row">
-            <div className="movie-card__reviews-col">
-              <div className="review">
-                <blockquote className="review__quote">
-                  <p className="review__text">Discerning travellers and Wes Anderson fans will luxuriate in the glorious
-                    Mittel-European kitsch of one of the director&lsquo;s funniest and most exquisitely designed movies in
-                    years.</p>
-
-                  <footer className="review__details">
-                    <cite className="review__author">Kate Muir</cite>
-                    <time className="review__date" dateTime="2016-12-24">December 24, 2016</time>
-                  </footer>
-                </blockquote>
-
-                <div className="review__rating">8,9</div>
-              </div>
-
-              <div className="review">
-                <blockquote className="review__quote">
-                  <p className="review__text">Anderson&lsquo;s films are too precious for some, but for those of us willing to
-                    lose ourselves in them, they&lsquo;re a delight. &quot;The Grand Budapest Hotel&quot; is no different, except that
-                    he has added a hint of gravitas to the mix, improving the recipe.</p>
-
-                  <footer className="review__details">
-                    <cite className="review__author">Bill Goodykoontz</cite>
-                    <time className="review__date" dateTime="2015-11-18">November 18, 2015</time>
-                  </footer>
-                </blockquote>
-
-                <div className="review__rating">8,0</div>
-              </div>
-
-              <div className="review">
-                <blockquote className="review__quote">
-                  <p className="review__text">I didn&lsquo;t find it amusing, and while I can appreciate the creativity, it&lsquo;s
-                    an hour and 40 minutes I wish I could take back.</p>
-
-                  <footer className="review__details">
-                    <cite className="review__author">Amanda Greever</cite>
-                    <time className="review__date" dateTime="2015-11-18">November 18, 2015</time>
-                  </footer>
-                </blockquote>
-
-                <div className="review__rating">8,0</div>
-              </div>
-            </div>
-            <div className="movie-card__reviews-col">
-              <div className="review">
-                <blockquote className="review__quote">
-                  <p className="review__text">The mannered, madcap proceedings are often delightful, occasionally silly,
-                    and here and there, gruesome and/or heartbreaking.</p>
-
-                  <footer className="review__details">
-                    <cite className="review__author">Matthew Lickona</cite>
-                    <time className="review__date" dateTime="2016-12-20">December 20, 2016</time>
-                  </footer>
-                </blockquote>
-
-                <div className="review__rating">7,2</div>
-              </div>
-
-              <div className="review">
-                <blockquote className="review__quote">
-                  <p className="review__text">It is certainly a magical and childlike way of storytelling, even if the
-                    content is a little more adult.</p>
-
-                  <footer className="review__details">
-                    <cite className="review__author">Paula Fleri-Soler</cite>
-                    <time className="review__date" dateTime="2016-12-20">December 20, 2016</time>
-                  </footer>
-                </blockquote>
-
-                <div className="review__rating">7,6</div>
-              </div>
-
-              <div className="review">
-                <blockquote className="review__quote">
-                  <p className="review__text">It is certainly a magical and childlike way of storytelling, even if the
-                    content is a little more adult.</p>
-
-                  <footer className="review__details">
-                    <cite className="review__author">Paula Fleri-Soler</cite>
-                    <time className="review__date" dateTime="2016-12-20">December 20, 2016</time>
-                  </footer>
-                </blockquote>
-
-                <div className="review__rating">7,0</div>
-              </div>
-            </div>
-          </div>
-        </>;
+        return <TabReviews />;
     }
 
-    return <>
-      <div className="movie-rating">
-        <div className="movie-rating__score">{film.rating}</div>
-        <p className="movie-rating__meta">
-          <span className="movie-rating__level">{getRatingString(film.rating)}</span>
-          <span className="movie-rating__count">{film.scores_count} ratings</span>
-        </p>
-      </div>
-
-      <div className="movie-card__text">
-        <p>In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge
-          Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave&apos;s friend and protege.</p>
-
-        <p>Gustave prides himself on providing first-class service to the hotel&apos;s guests, including satisfying the
-          sexual needs of the many elderly women who stay there. When one of Gustave&apos;s lovers dies mysteriously,
-          Gustave finds himself the recipient of a priceless painting and the chief suspect in her murder.</p>
-
-        <p className="movie-card__director"><strong>Director: {film.director}</strong></p>
-
-        <p className="movie-card__starring"><strong>Starring: {film.starring.map((elem) => (`${elem}, `))}
-          and other</strong></p>
-      </div>
-    </>;
+    return <TabOverview film={film} />;
   };
 
   return <>
